pyproject.toml
[tool.poetry]
name = "bad_translation_bot"
version = "0.1.0"
description = "Bad Translation Discord Bot"
authors = ["Carson Mullins <septem151@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-translate = "^3.11.1"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
validators = "^0.20.0"
py-cord = "^2.4.1"
emoji = "^2.4.0"
pronouncing = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
isort = "^5.12.0"
black = "^23.3.0"
mypy = "^1.3.0"
pylint = "^2.17.4"
types-requests = "^2.31.0.1"
types-emoji = "^2.1.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
warn_unreachable = true
warn_redundant_casts = true
warn_unused_ignores = true
show_column_numbers = true
check_untyped_defs = true
follow_imports = "silent"
exclude = "doc/"

[[tool.mypy.overrides]]
module = [
  "google.cloud.translate_v2",
  "validators",
  "pronouncing",
  "pytest"
]
ignore_missing_imports = true

[tool.isort]
profile = "black"

[tool.pylint.messages_control]
disable = "C0114, C0115, C0116, W0511"

[tool.pylint.design]
max-args = "6"

[tool.pylint.format]
max-line-length = "88"
